Only set comint-use-prompt-regexp to t for comint-send-input
We want it to be non-nil to remove things like "o1 = " from the input
(see 6449756), but we want it to be nil otherwise so that the output
and input fields are separated properly.  In particular, so that C-a
and C-0 C-k move the point to the end of input prompt and not the very
beginning of the line.

Closes: #24