Add menus to M2-mode and M2-comint-mode
These menus may be useful to users for discover what functions are
available in the two Macaulay2 major modes and to quickly learn their keybindings.

The menus include all functions marked "interactive".  Functions that
have a corresponding keybinding in M2-mode only are only displayed in
the M2-mode menu, and similar for M2-comint-mode.  Functions that
have a corresponding keybinding in both modes are displayed in both
menus.

There are some exceptions to this:

* M2 is bound to f12 in both modes, but really only makes sense in the
  M2-mode menu.
* M2-send-to-program is bound to f11 in both modes, but makes the most
  sense in the M2-mode menu, as
  M2-send-to-program-or-jump-to-source-code serves a similar purpose
  in M2-comint-mode.
* M2-dynamic-complete-symbol is marked obsolete, so it is omitted.
* M2-electric-right-brace has its keybinding commented out, but we
  include it for completeness.
* M2-electric-tab is set to indent-line-function for both modes, but
  TAB is bound to completion-at-point in M2-comint-mode, so we only
  include it in the M2-mode menu.