Deprecate M2-send-input in favor of comint-send-input
The only difference is that we were forcing comint-use-prompt-regexp
to be t, but sending code to Macaulay2 works just fine whether it's t
or nil.  The only difference is whether we send just a single line or
the entire input field to the M2 process when scrolling up to previous
inputs and pressing return.  Let's let the user choose the desired
behavior by setting comint-use-prompt-regexp to t if they choose.  (It
default to nil globally).