Update filename-matching regular expression
In particular, compilation mode will only recognize paths containing
alphanumeric characters, "/", ".", "_", and "-" (or, as before,
paths enclosed in double quotes which will happen when they contain a
space).  This should catch the vast majority of paths so we don't need
to worry about creating a specific blacklist for things like spaces,
"[", "{", "|", etc.  Initial hyphens are disallowed.

Closes: #71