Adjust indentation when lines begin with right parens
If the first character in the current line is a right paren, then we
decrease the indentation.  But to offset this, we need to increase
the indentation when the first character in the previous line is a
right paren.

Also merge the otherwise unused function M2-next-line-indent-amount
into M2-this-line-indent-amount.