Close any unclosed strings or comments from the output
We add a new function, M2-comint-fix-unclosed, that is called every
time we receive new output from the M2 process.  It detects if we're
still inside a string or comment, and if so, invisibly inserts the
corresponding string right before the input prompt to terminate it and
reset the syntax highlighting.